requirements.txt
aiogram>=3.1.1
python-dotenv>=1.0.0
web3==6.15.1
dataclasses-json==0.6.4
aiohttp_cors==0.7.0